+++ src/components/plot/outline-popover.tsx
@@ -1,30 +1,13 @@
 import { BaseCol } from '@/layout/base-col'
-import { VCenterRow } from '@/layout/v-center-row'
-import { Button, type IButtonProps } from '@/themed/v2/button'
+import { type IButtonProps } from '@/themed/v2/button'
 
-import {
-  addAlphaToHex,
-  COLOR_BLACK,
-  COLOR_TRANSPARENT,
-  hexToRgba,
-  hexColorWithoutAlpha as removeAlphaFromHex,
-  rgba2hex,
-  textColorShouldBeDark,
-  type IRGBA,
-} from '@/lib/color/color'
+import { hexToRgba } from '@/lib/color/color'
 import { cn } from '@/lib/shadcn-utils'
 
 import { FOCUS_RING_CLS } from '@/theme'
 
-import { CircleSlash, SquarePen } from 'lucide-react'
-import { useState } from 'react'
-import {
-  HexAlphaColorPicker,
-  HexColorInput,
-  HexColorPicker,
-} from 'react-colorful'
-import { CheckPropRow } from '../dialogs/check-prop-row'
-import { NumericalInput } from '../shadcn/ui/themed/numerical-input'
+import { CircleSlash, Palette, SquarePen } from 'lucide-react'
+import { useDialogs } from '../dialogs/dialogs'
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -36,7 +19,6 @@ import {
   DropdownMenuTrigger,
   MenuSeparator,
 } from '../shadcn/ui/themed/v2/dropdown-menu'
-import { inputVariants } from '../shadcn/ui/themed/v2/input'
 import { IColorPickerProps } from './color-picker-popover'
 import { THEME_COLOR_GRID } from './theme'
 
@@ -141,24 +123,6 @@ export function OutlineButton({
     return null
   }
 
-  const color = colors[0]!
-
-  const lightMode = textColorShouldBeDark(color.color)
-
-  // ensure alpha is maintained
-  const _color = color.keepAlphaChannel
-    ? color.color
-    : addAlphaToHex(color.color, color.opacity ?? 1)
-
-  const border =
-    (color.autoBorder && lightMode) || _color === COLOR_TRANSPARENT
-      ? 'border-border'
-      : (color.defaultBorderColor ?? 'border-foreground')
-
-  const textColor = lightMode
-    ? 'text-foreground stroke-foreground'
-    : 'text-background stroke-background'
-
   // if (!ariaLabel) {
   //   ariaLabel = tooltip
   // }
@@ -189,39 +153,35 @@ export function OutlinePopover({
   colors,
 
   align = 'start',
-  open,
+  //open,
   onOpenChanged,
 
   children,
   ...props
 }: IOutlineButtonProps) {
-  const [_open, setOpen] = useState(false)
-  const o = open ?? _open
+  //const [_open, setOpen] = useState(false)
+  //const o = open ?? _open
+
+  const { open: openDialog } = useDialogs()
 
   if (!colors || colors.length === 0) {
     return null
   }
-  //const [_show, setShow] = useState(showColor)
 
-  // function _onColorChange(color: string) {
-  //   const rgba = hexToRgba(color)
-
-  //   onColorChange?.(keepAlphaChannel ? color : color.slice(0, 7), rgba[3])
-  // }
-
-  function _openChanged(open: boolean) {
-    setOpen(open)
-    onOpenChanged?.(open)
-  }
+  // function _openChanged(open: boolean) {
+  //  setOpen(open)
+  //   onOpenChanged?.(open)
+  //}
 
   // if (tooltip) {
   //   button = <Tooltip content={tooltip}>{button}</Tooltip>
   // }
 
   const color0 = colors[0]!
+  const opacity = color0.opacity ?? 1
 
   return (
-    <DropdownMenu open={o} onOpenChange={(v) => _openChanged(v)}>
+    <DropdownMenu>
       {children}
 
       <DropdownMenuContent
@@ -237,8 +197,6 @@ export function OutlinePopover({
           <div className="grid grid-cols-10 items-center gap-x-2">
             {THEME_COLOR_GRID.map((group, gi) => (
               <BaseCol key={gi} className="gap-y-1">
-                {/* <span className="text-xs font-medium">{group.name}</span> */}
-
                 {group.colors.map((presetColor) => {
                   const prgb = hexToRgba(presetColor)
                   const ps = prgb[0] + prgb[1] + prgb[2]
@@ -256,9 +214,7 @@ export function OutlinePopover({
                       onClick={() =>
                         color0.onColorChange?.({
                           color: presetColor,
-                          opacity: color0.opacity ?? 1,
-                          width: color0.width ?? 1,
-                          dasharray: color0.dasharray ?? '0',
+                          opacity,
                         })
                       }
                       tabIndex={0}
@@ -274,38 +230,68 @@ export function OutlinePopover({
         <BaseCol className="gap-y-2 mb-2 mx-1.5">
           <span className="text-xs font-semibold">Standard Colors</span>
           <div className="grid grid-cols-10 items-center gap-x-2 gap-y-1">
-            {PRESET_COLORS.map((presetColor, gi) => {
-              const prgb = hexToRgba(presetColor)
-              const ps = prgb[0] + prgb[1] + prgb[2]
+            {[...PRESET_COLORS.slice(0, 10), color0.color].map(
+              (presetColor, pi) => {
+                const prgb = hexToRgba(presetColor)
+                const ps = prgb[0] + prgb[1] + prgb[2]
 
-              return (
-                <button
-                  key={presetColor}
-                  className={cn(
-                    'w-4.5 rounded-full aspect-square border hover:scale-125 focus-visible:scale-125 transition-transform duration-300',
-                    color0.autoBorder && ps >= 750 && 'border-border',
-                    !color0.autoBorder ||
-                      (ps < 750 && 'border-transparent hover:border-white')
-                  )}
-                  style={{ background: presetColor }}
-                  onClick={() =>
-                    color0.onColorChange?.({
-                      color: presetColor,
-                      opacity: color0.opacity ?? 1,
-                      width: color0.width ?? 1,
-                      dasharray: color0.dasharray ?? '0',
-                    })
-                  }
-                  tabIndex={0}
-                />
-              )
-            })}
+                return (
+                  <button
+                    key={`${presetColor}-${pi}`}
+                    className={cn(
+                      'w-4.5 rounded-full aspect-square border hover:scale-125 focus-visible:scale-125 transition-transform duration-300',
+                      color0.autoBorder && ps >= 750 && 'border-border',
+                      !color0.autoBorder ||
+                        (ps < 750 && 'border-transparent hover:border-white')
+                    )}
+                    style={{ background: presetColor }}
+                    onClick={() =>
+                      color0.onColorChange?.({
+                        color: presetColor,
+                        opacity,
+                      })
+                    }
+                    tabIndex={0}
+                  />
+                )
+              }
+            )}
           </div>
         </BaseCol>
 
-        <DropdownMenuItem>
+        <DropdownMenuItem
+          onClick={() => {
+            color0.onColorChange?.({
+              color: color0.color,
+              opacity,
+              show: false,
+            })
+          }}
+        >
           <CircleSlash size={20} strokeWidth={1.5} />
           <span>No Outline</span>
+        </DropdownMenuItem>
+        <MenuSeparator />
+
+        <DropdownMenuItem
+          onClick={() => {
+            openDialog({
+              type: 'color',
+              payload: {
+                color: color0,
+
+                callback: (color) => {
+                  color0.onColorChange?.({
+                    color: color.color,
+                    opacity: color.opacity,
+                  })
+                },
+              },
+            })
+          }}
+        >
+          <Palette size={20} strokeWidth={1.5} />
+          <span>More Colors...</span>
         </DropdownMenuItem>
         <MenuSeparator />
 
@@ -313,13 +299,37 @@ export function OutlinePopover({
           <DropdownMenuSubTrigger>Weight</DropdownMenuSubTrigger>
           <DropdownMenuPortal>
             <DropdownMenuSubContent>
-              <DropdownMenuItem onClick={() => {}}>
+              <DropdownMenuItem
+                onClick={() => {
+                  color0.onColorChange?.({
+                    color: color0.color,
+                    opacity,
+                    width: 0.5,
+                  })
+                }}
+              >
                 <span>0.5 pt</span>
               </DropdownMenuItem>
-              <DropdownMenuItem onClick={() => {}}>
+              <DropdownMenuItem
+                onClick={() => {
+                  color0.onColorChange?.({
+                    color: color0.color,
+                    opacity,
+                    width: 1,
+                  })
+                }}
+              >
                 <span>1 pt</span>
               </DropdownMenuItem>
-              <DropdownMenuItem onClick={() => {}}>
+              <DropdownMenuItem
+                onClick={() => {
+                  color0.onColorChange?.({
+                    color: color0.color,
+                    opacity,
+                    width: 2,
+                  })
+                }}
+              >
                 <span>2 pt</span>
               </DropdownMenuItem>
             </DropdownMenuSubContent>
@@ -329,15 +339,39 @@ export function OutlinePopover({
           <DropdownMenuSubTrigger>Dashes</DropdownMenuSubTrigger>
           <DropdownMenuPortal>
             <DropdownMenuSubContent>
-              <DropdownMenuItem onClick={() => {}}>
-                <span className="w-3/5 h-0.5 bg-foreground  " />
+              <DropdownMenuItem
+                onClick={() => {
+                  color0.onColorChange?.({
+                    color: color0.color,
+                    opacity,
+                    dasharray: '0',
+                  })
+                }}
+              >
+                <span className="w-3/5 h-0.5 bg-foreground" />
                 <span>Solid</span>
               </DropdownMenuItem>
-              <DropdownMenuItem onClick={() => {}}>
+              <DropdownMenuItem
+                onClick={() => {
+                  color0.onColorChange?.({
+                    color: color0.color,
+                    opacity,
+                    dasharray: '2',
+                  })
+                }}
+              >
                 <span className="w-3/5 h-0 border-t-2 border-foreground border-dotted" />
                 <span>Round Dot</span>
               </DropdownMenuItem>
-              <DropdownMenuItem onClick={() => {}}>
+              <DropdownMenuItem
+                onClick={() => {
+                  color0.onColorChange?.({
+                    color: color0.color,
+                    opacity,
+                    dasharray: '8',
+                  })
+                }}
+              >
                 <span className="w-3/5 h-0 border-t-2 border-foreground border-dashed" />
                 <span>Dash</span>
               </DropdownMenuItem>
@@ -346,259 +380,5 @@ export function OutlinePopover({
         </DropdownMenuSub>
       </DropdownMenuContent>
     </DropdownMenu>
-  )
-}
-
-export function OutlineUI({
-  title,
-  color,
-  opacity,
-  defaultColor,
-  onColorChange,
-  autoBorder = true,
-  allowNoColor = false,
-  showPresets = true,
-  showColor = true,
-  showRgb = false,
-  onShowColor,
-  width,
-  dasharray,
-
-  // we default to removing the alpha channel from the hex color
-  // itself and instead returning it via the alpha parameter. This
-  // is because this control is typically used for setting colors
-  // that will be used in SVG components and to play nicely, we use
-  // colors without the alpha channel and instead use fillOpacity and
-  // strokeOpacity to set the color opacity on the svg element.
-  keepAlphaChannel = false,
-}: IColorPickerProps) {
-  const _width = width ?? 1
-  const _opacity = opacity ?? 1
-
-  const [_show, setShow] = useState(showColor)
-
-  const _colorWithAlpha = addAlphaToHex(color, _opacity)
-
-  const _rgba = hexToRgba(_colorWithAlpha)
-
-  function _onColorChange(
-    color: string,
-    opacity: number = 1,
-    width: number = 1,
-    dasharray: string = '0'
-  ) {
-    color = addAlphaToHex(color, opacity)
-
-    const rgba = hexToRgba(color)
-
-    onColorChange?.({
-      color: keepAlphaChannel ? color : removeAlphaFromHex(color),
-      opacity: rgba[3],
-      width,
-      dasharray,
-    })
-  }
-
-  return (
-    <BaseCol className="color-picker gap-y-2">
-      {title && <span className="text-xs font-bold">{title}</span>}
-
-      {opacity !== undefined ? (
-        <HexAlphaColorPicker
-          color={_colorWithAlpha ?? COLOR_BLACK}
-          onChange={(v) => {
-            const a = hexToRgba(v)[3]
-
-            _onColorChange(v, a, width ?? 1, dasharray ?? '0')
-          }}
-        />
-      ) : (
-        <HexColorPicker
-          color={color ?? COLOR_BLACK}
-          onChange={_onColorChange}
-        />
-      )}
-
-      <VCenterRow className="gap-x-3 ">
-        <VCenterRow className="gap-x-1.5">
-          <span>Hex</span>
-          <HexColorInput
-            id="hex"
-            color={color?.toUpperCase() ?? COLOR_BLACK}
-            alpha={true}
-            prefixed={true}
-            onChange={_onColorChange}
-            className={inputVariants({
-              variant: 'alt',
-              className: 'w-20 h-9',
-            })}
-          />
-        </VCenterRow>
-
-        {opacity !== undefined && (
-          <VCenterRow className="gap-x-1.5">
-            <span>A</span>
-            <NumericalInput
-              dp={2}
-              value={opacity}
-              className="w-16"
-              onNumChanged={(v) => {
-                const a = Math.max(0, Math.min(1, v))
-
-                // if keepAlphaChannel is true, we need to reconstruct the color
-                // to include the new alpha, otherwise strip the alpha from the color
-                _onColorChange(
-                  keepAlphaChannel
-                    ? addAlphaToHex(color, a)
-                    : removeAlphaFromHex(color),
-                  a,
-                  _width
-                )
-              }}
-              limit={[0, 1]}
-              step={0.1}
-              variant="alt"
-            />
-          </VCenterRow>
-        )}
-      </VCenterRow>
-      {showRgb && (
-        <VCenterRow className="gap-x-3">
-          <VCenterRow className="gap-x-1.5">
-            <span>R</span>
-            <NumericalInput
-              value={_rgba[0]}
-              onNumChange={(v) => {
-                const rgba: IRGBA = [..._rgba]
-                rgba[0] = v
-                _onColorChange(rgba2hex(rgba), rgba[3], _width)
-              }}
-              min={0}
-              max={255}
-              step={1}
-              variant="alt"
-              className="w-14"
-            />
-          </VCenterRow>
-          <VCenterRow className="gap-x-1.5">
-            <span>G</span>
-            <NumericalInput
-              value={_rgba[1]}
-              onNumChange={(v) => {
-                const rgba: IRGBA = [..._rgba]
-                rgba[1] = v
-                _onColorChange(rgba2hex(rgba), rgba[3], _width)
-              }}
-              min={0}
-              max={255}
-              step={1}
-              variant="alt"
-              className="w-14"
-            />
-          </VCenterRow>
-          <VCenterRow className="gap-x-1.5">
-            <span>B</span>
-            <NumericalInput
-              value={_rgba[2]}
-              onNumChange={(v) => {
-                const rgba: IRGBA = [..._rgba]
-                rgba[2] = v
-                _onColorChange(rgba2hex(rgba), rgba[3], _width)
-              }}
-              min={0}
-              max={255}
-              step={1}
-              variant="alt"
-              className="w-14"
-            />
-          </VCenterRow>
-        </VCenterRow>
-      )}
-
-      {showPresets && (
-        <VCenterRow className="gap-x-2">
-          <div className="grid grid-cols-10 gap-1 items-center justify-center">
-            {PRESET_COLORS.map((presetColor) => {
-              const prgb = hexToRgba(presetColor)
-              const ps = prgb[0] + prgb[1] + prgb[2]
-
-              return (
-                <button
-                  key={presetColor}
-                  className={cn(
-                    'w-5.5 aspect-square border hover:scale-125 focus-visible:scale-125 rounded-xs transition-transform duration-300',
-                    autoBorder && ps >= 750 && 'border-border',
-                    !autoBorder ||
-                      (ps < 750 && 'border-transparent hover:border-white')
-                  )}
-                  style={{ background: presetColor }}
-                  onClick={() => _onColorChange(presetColor, _opacity, _width)}
-                  tabIndex={0}
-                />
-              )
-            })}
-          </div>
-        </VCenterRow>
-      )}
-
-      {/* {(allowNoColor || defaultColor) && <MenuSeparator />} */}
-
-      {allowNoColor && (
-        <Button
-          variant="flat"
-          className="w-full"
-          justify="start"
-          rounded="md"
-          onClick={() => _onColorChange(COLOR_TRANSPARENT)}
-        >
-          <span className="relative aspect-square w-5 border border-border bg-background rounded-xs">
-            <span className="absolute left-0 w-full bg-red-400 h-px top-1/2 -translate-y-1/2 -rotate-45" />
-          </span>
-
-          <span>No color</span>
-        </Button>
-      )}
-
-      {width !== undefined && (
-        <VCenterRow className="gap-x-2">
-          <span className="w-10">Width</span>
-          <NumericalInput
-            value={width}
-            onNumChange={(v) => _onColorChange(color, _opacity, v)}
-            min={0}
-            step={1}
-            max={1000}
-            placeholder="Width"
-          />
-        </VCenterRow>
-      )}
-
-      {onShowColor && (
-        <CheckPropRow
-          title="Show"
-          checked={_show}
-          onCheckedChange={(v) => {
-            setShow(v)
-            onShowColor?.(v)
-          }}
-        />
-      )}
-
-      {defaultColor && (
-        <Button
-          variant="flat"
-          className="w-full"
-          justify="start"
-          rounded="md"
-          onClick={() => _onColorChange(defaultColor)}
-        >
-          <span
-            className="aspect-square w-5 border border-border rounded-xs"
-            style={{ backgroundColor: defaultColor }}
-          />
-          <span>Reset color</span>
-        </Button>
-      )}
-    </BaseCol>
   )
 }